requirements.txt
Pillow
inky
font_hanken_grotesk